Cleanup some compiler warnings.

+++ fcheck.c
@@ -24,7 +24,7 @@
 
 #include "funcs.c"
 
-#define VARSNUM		((int)(sizeof(Vars) / sizeof(struct avar)))
+#define VARSNUM		((sizeof(Vars) / sizeof(struct avar)))
 
 void Dbg(char *fmt, ...) {}
 
@@ -33,13 +33,13 @@ int main(int argc, char *argv[])
 	int i, err = 0;
 
 	if (NUMVARS != VARSNUM) {
-		printf("Mismatch in NUMVARS and VARNUM %d:%d\n",
+		printf("Mismatch in NUMVARS and VARNUM %d:%ld\n",
 		       NUMVARS, VARSNUM);
 		err = 1;
 	}
 
 	if (N_KEYS != NUM_SPECIAL) {
-		printf("Mismatch N_KEYS %d NUMKEYS %d\n",
+		printf("Mismatch N_KEYS %ld NUMKEYS %d\n",
 		       N_KEYS, NUM_SPECIAL);
 		err = 1;
 	}
